don't rerun the example - too large of a memory footprint for travis
